Encapsulate common nginx conf
A build.sh script will be added that allows variables specified in
nginx conf files to be replaced. The script is run by the grunt pipeline
in order to generate the test set of nginx files, and the script is
run by the post-receive production hook to generate the production set
of nginx files.